+++ .gitignore
@@ -5,3 +5,4 @@
 /src/aaio.egg-info
 /dist/
 /site/
+main.py
