+++ pyproject.toml
@@ -1,10 +1,13 @@
 [build-system]
-requires = ["setuptools>=61.0"]
-build-backend = "setuptools.build_meta"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/aaio"]
 
 [project]
 name = "aaio"
-version = "1.3.1"
+version = "1.4.0"
 description = "Async AAIO api wrapper for python"
 readme = "README.md"
 authors = [{ name = "kewldan", email = "kewldanil1@gmail.com" }]
