+++ src/aaio/client.py
@@ -1,31 +1,13 @@
 import hashlib
-from typing import Optional, List
+import logging
+from typing import List, Optional
 from urllib.parse import urlencode
 
 import aiohttp
 
-from aaio.exceptions.aaio_bad_request import AAIOBadRequest
-from aaio.types.balance import Balance
-from aaio.types.create_payoff import CreatePayoff
-from aaio.types.payment_info import PaymentInfo
-from aaio.types.payment_methods import PaymentMethods
-from aaio.types.payment_webhook import PaymentWebhookData
-from aaio.types.payoff_info import PayoffInfo
-from aaio.types.payoff_methods import PayoffMethods
-from aaio.types.payoff_rates import PayoffRates
-from aaio.types.payoff_sbp_banks import PayoffSbpBanks
-from aaio.types.payoff_webhook import PayoffWebhookData
-
-
-async def create_invoice(payment_url: str):
-    async with aiohttp.ClientSession() as session:
-        async with session.get(payment_url) as request:
-            return request.url
-
-
-def is_valid_payoff_webhook(data: PayoffWebhookData, secret_key: str) -> bool:
-    return hashlib.sha256(
-        f'{data.id}:{secret_key}:{data.amount_down}'.encode()).hexdigest() == data.sign
+from .exceptions import AAIOBadRequest
+from .models import PaymentInfo, Balance, CreatePayoff, PayoffSbpBanks, PayoffInfo, PayoffRates, PayoffMethods, \
+    PaymentMethods, PaymentWebhookData
 
 
 class AAIO:
@@ -35,7 +17,7 @@ class AAIO:
 
     API for https://aaio.so/
     """
-    
+
     def __init__(self, merchant_id: str, secret_1: str, secret_2: str, api_key: str, default_currency: str = 'RUB',
                  base_url: str = 'https://aaio.so'):
         """
@@ -49,14 +31,14 @@ class AAIO:
             default_currency: If not set - RUB, but can be overwritten for each request (Optional)
             base_url: Base URL for requests (Optional)
         """
-        
+
         self._default_currency = default_currency
         self._merchant_id = merchant_id
         self._api_key = api_key
         self._secret_1 = secret_1
         self._secret_2 = secret_2
         self._base_url = base_url
-    
+
     def __generate_sign(self, amount: float, order_id: str, currency: str) -> str:
         """
         Generates sign for payment creation
@@ -70,7 +52,7 @@ class AAIO:
         Returns: SHA-256 sign
 
         """
-        
+
         params = f':'.join([
             self._merchant_id,
             str(amount),
@@ -80,13 +62,31 @@ class AAIO:
         ])
         sign = hashlib.sha256(params.encode('utf-8')).hexdigest()
         return sign
-    
+
+    def __generate_payment_params(self, amount: float, order_id: str, description: str = None, method: str = None,
+                                  email: str = None,
+                                  referral: str = None, us_key: str = None, currency: str = None,
+                                  language: str = 'ru'):
+        return {
+            'merchant_id': self._merchant_id,
+            'amount': amount,
+            'currency': currency,
+            'order_id': order_id,
+            'sign': self.__generate_sign(amount, order_id, currency or self._default_currency),
+            'desc': description,
+            'lang': language,
+            'method': method,
+            'email': email,
+            'referral': referral,
+            'us_key': us_key
+        }
+
     def create_payment(self, amount: float, order_id: str, description: str = None, method: str = None,
                        email: str = None,
                        referral: str = None, us_key: str = None, currency: str = None,
                        language: str = 'ru') -> str:
         """
-        Creates payment URL (Not a request)
+        Generates payment URL (DEPRECATED)
         See https://wiki.aaio.so/priem-platezhei/sozdanie-zakaza for more detailed information
 
         Args:
@@ -103,31 +103,20 @@ class AAIO:
         Returns: Payment URL
 
         """
-        
-        if not currency:
-            currency = self._default_currency
-        params = {
-            'merchant_id': self._merchant_id,
-            'amount': amount,
-            'currency': currency,
-            'order_id': order_id,
-            'sign': self.__generate_sign(amount, order_id, currency),
-            'desc': description,
-            'lang': language,
-            'method': method,
-            'email': email,
-            'referral': referral,
-            'us_key': us_key
-        }
-        
+
+        logging.warning('This method is deprecated, consider using create_pay method instead')
+
+        params = self.__generate_payment_params(amount, order_id, description, method, email, referral, us_key,
+                                                currency, language)
+
         return f'{self._base_url}/merchant/pay?' + urlencode({k: v for k, v in params.items() if v is not None})
-    
+
     async def get_pay_url(self, amount: float, order_id: str, description: str = None, method: str = None,
                           email: str = None,
                           referral: str = None, us_key: str = None, currency: str = None,
                           language: str = 'ru') -> dict:
         """
-        Creates payment URL, recomended method
+        Creates payment URL
         See https://wiki.aaio.so/priem-platezhei/sozdanie-zakaza-zaprosom-rekomenduem for more detailed information
 
         Args:
@@ -143,32 +132,19 @@ class AAIO:
 
         Returns: dict {"type": "...", "url": "https://......"}
         """
-        
-        if not currency:
-            currency = self._default_currency
-        params = {
-            'merchant_id': self._merchant_id,
-            'amount': amount,
-            'currency': currency,
-            'order_id': order_id,
-            'sign': self.__generate_sign(amount, order_id, currency),
-            'desc': description,
-            'lang': language,
-            'method': method,
-            'email': email,
-            'referral': referral,
-            'us_key': us_key
-        }
-        
+
+        params = self.__generate_payment_params(amount, order_id, description, method, email, referral, us_key,
+                                                currency, language)
+
         response = await self.__create_request('/merchant/get_pay_url', params)
-        
+
         return response
-    
+
     async def get_ips(self) -> List[str]:
         response = await self.__create_request('/api/public/ips')
-        
+
         return response['list']
-    
+
     async def get_payment_info(self, order_id: str) -> PaymentInfo:
         """
         Creates a request for get payment information
@@ -180,16 +156,16 @@ class AAIO:
         Returns: Model from response JSON
 
         """
-        
+
         params = {
             'merchant_id': self._merchant_id,
             'order_id': order_id
         }
-        
+
         response = await self.__create_request('/api/info-pay', params)
-        
+
         return PaymentInfo(**response)
-    
+
     async def get_balances(self) -> Balance:
         """
         Creates a request for get balances of user
@@ -197,11 +173,11 @@ class AAIO:
 
         Returns: Model from response JSON
         """
-        
+
         response = await self.__create_request('/api/balance')
-        
+
         return Balance(**response)
-    
+
     async def create_payoff(self, method: str, amount: float, wallet: str, payoff_id: str = '',
                             commission_type: int = 0) -> CreatePayoff:
         """
@@ -218,7 +194,7 @@ class AAIO:
         Returns: Model from response JSON
 
         """
-        
+
         params = {
             'my_id': payoff_id,
             'method': method,
@@ -226,11 +202,11 @@ class AAIO:
             'wallet': wallet,
             'commission_type': commission_type
         }
-        
+
         response = await self.__create_request('/api/create-payoff', params)
-        
+
         return CreatePayoff(**response)
-    
+
     async def get_payoff_sbp_banks(self) -> PayoffSbpBanks:
         """
         Returns a list of available banks for payoff
@@ -238,9 +214,9 @@ class AAIO:
         Returns: list of banks
         """
         response = await self.__create_request('/api/sbp-banks-payoff')
-        
+
         return PayoffSbpBanks(**response)
-    
+
     async def get_payoff_info(self, payoff_id: str = None, aaio_id: str = None) -> PayoffInfo:
         """
         Creates a request for get payoff information
@@ -255,16 +231,16 @@ class AAIO:
         Returns: Model from response JSON
 
         """
-        
+
         params = {
             'my_id': payoff_id,
             'id': aaio_id
         }
-        
+
         response = await self.__create_request('/api/info-payoff', params)
-        
+
         return PayoffInfo(**response)
-    
+
     async def get_payoff_rates(self) -> PayoffRates:
         """
         Creates a request for get rates for payoff
@@ -273,11 +249,11 @@ class AAIO:
         Returns: Model from response JSON
 
         """
-        
+
         response = await self.__create_request('/api/rates-payoff')
-        
+
         return PayoffRates(**response)
-    
+
     async def get_payoff_methods(self) -> PayoffMethods:
         """
         Creates a request for get available payoff methods
@@ -286,11 +262,11 @@ class AAIO:
         Returns: Model from response JSON
 
         """
-        
+
         response = await self.__create_request('/api/methods-payoff')
-        
+
         return PayoffMethods(**response)
-    
+
     async def get_payment_methods(self) -> PaymentMethods:
         """
         Creates a request for get available payment methods
@@ -299,15 +275,15 @@ class AAIO:
         Returns: Model from response JSON
 
         """
-        
+
         params = {
             'merchant_id': self._merchant_id
         }
-        
+
         response = await self.__create_request('/api/methods-pay', params)
-        
+
         return PaymentMethods(**response)
-    
+
     async def __create_request(self, uri: str, params: dict = None) -> Optional[dict]:
         """
         Creates a request to base URL and adds URI
@@ -319,16 +295,16 @@ class AAIO:
         Returns: Model from response JSON
 
         """
-        
+
         if params is None:
             params = {}
-        
+
         headers = {
             'Accept': 'application/json',
             'Content-Type': 'application/x-www-form-urlencoded',
             'X-Api-Key': self._api_key
         }
-        
+
         async with aiohttp.ClientSession(self._base_url) as session:
             async with session.post(uri, headers=headers,
                                     data={k: v for k, v in params.items() if v is not None}) as r:
@@ -337,7 +313,7 @@ class AAIO:
                     return response
                 else:
                     raise AAIOBadRequest(response['code'], response['message'])
-    
+
     def is_valid_payment_webhook(self, data: PaymentWebhookData) -> bool:
         return hashlib.sha256(
             f'{self._merchant_id}:{data.amount}:{self._secret_2}:{data.order_id}'.encode()).hexdigest() == data.sign
