+++ src/aaio/exceptions/__init__.py
@@ -1,0 +1,1 @@
+from .bad_request import AAIOBadRequest
