+++ src/aaio/exceptions/bad_request.py
@@ -1,4 +1,4 @@
-class AAIOBadRequest(BaseException):
+class AAIOBadRequest(Exception):
     def __init__(self, code: int, message: str) -> None:
         self.code = code
         self.message = message
