+++ src/aaio/models/payoff_methods.py
