+++ src/aaio/models/payoff_rates.py
